Load 3.3.488 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1043 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/array.c
@@ -2733,7 +2733,7 @@ f_transpose_array (svalue_t *sp)
         z = y->u.vec->item;
 
         c = b;
-        if (VEC_SIZE(y->u.vec) < (size_t)b
+        if (VEC_SIZE(y->u.vec) < b
          && !(c = (mp_int)VEC_SIZE(y->u.vec)) )
                 continue;
